pyproject.toml
[tool.poetry]
name = "msc-thesis"
version = "0.1.0"
description = ""
authors = ["Noah Alting <noahalting@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
matplotlib = "^3.9.3"
laspy = "^2.5.4"
notebook = "^7.2.2"
ipykernel = "^6.29.5"
open3d = "^0.18.0"
rerun-sdk = "^0.20.2"
rerun-notebook = "^0.20.2"
sparqlwrapper = "^2.0.0"
geopandas = "^1.0.1"
shapely = "^2.0.6"
tqdm = "^4.67.1"
laszip = "^0.2.3"
rasterio = "^1.4.3"
pymeshlab = "^2023.12.post2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
